--- a/.gitignore
+++ b/.gitignore
@@ -1,2 +1,198 @@
-# Created by venv; see https://docs.python.org/3/library/venv.html
-*
+# build, distribute, and bins (+ python proto bindings)
+build
+build_*/
+.build_debug/*
+.build_release/*
+distribute/*
+*.testbin
+*.bin
+cmake_build
+.cmake_build
+cmake-build-debug
+gen
+*~
+.vs
+.DS_Store
+*.DS_Store
+TestResults/
+.idea/
+onnxruntime.egg-info
+nuget_root/
+.packages/
+.vscode
+*.code-workspace
+__pycache__
+onnxruntime_profile*.json
+/docs/python/inference/*.md
+/docs/python/inference/auto_examples/*
+/docs/python/inference/media/*
+/docs/python/inference/examples/*.onnx
+/docs/python/inference/examples/graph.*
+/docs/python/*_LICENSE
+/LICENSE.txt
+/csharp/**/obj/
+/csharp/**/bin/
+/csharp/Directory.Build.props
+docs/python/inference/*.onnx
+*.onnx
+onnxprofile_profile_test_*.json
+/csharp/packages
+/csharp/src/Microsoft.ML.OnnxRuntime/targets/**/*.targets
+/csharp/src/Microsoft.ML.OnnxRuntime/targets/**/*.props
+/csharp/**/*.vcxproj.user
+cmake/external/FeaturizersLibrary/
+# Java specific ignores
+java/.gradle
+java/hs_*.log
+/java/bin
+onnxruntime/python/version_info.py
+/orttraining/orttraining/eager/ort_aten.g.cpp
+/orttraining/orttraining/eager/ort_customops.g.cpp
+/csharp/**/packages
+# direnv, posh-direnv
+.envrc
+.psenvrc
+*.csproj.user
+# clangd
+.cache/
+compile_commands.json
+# Rust specific
+rust/**/target
+rust/**/Cargo.lock
+rust/onnxruntime/synset.txt
+
+# Python
+
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
+
+# C extensions
+*.so
+
+# Distribution / packaging
+.Python
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib64/
+parts/
+sdist/
+var/
+wheels/
+share/python-wheels/
+*.egg-info/
+.installed.cfg
+*.egg
+MANIFEST
+
+# PyInstaller
+#  Usually these files are written by a python script from a template
+#  before PyInstaller builds the exe, so as to inject date/other infos into it.
+*.manifest
+*.spec
+
+# Installer logs
+pip-log.txt
+pip-delete-this-directory.txt
+
+# Unit test / coverage reports
+htmlcov/
+.tox/
+.nox/
+.coverage
+.coverage.*
+.cache
+nosetests.xml
+coverage.xml
+*.cover
+*.py,cover
+.hypothesis/
+.pytest_cache/
+cover/
+
+# Translations
+*.mo
+*.pot
+
+# Django stuff:
+*.log
+local_settings.py
+db.sqlite3
+db.sqlite3-journal
+
+# Flask stuff:
+instance/
+.webassets-cache
+
+# Scrapy stuff:
+.scrapy
+
+# Sphinx documentation
+docs/_build/
+
+# PyBuilder
+.pybuilder/
+target/
+
+# Jupyter Notebook
+.ipynb_checkpoints
+
+# IPython
+profile_default/
+ipython_config.py
+
+# PEP 582; used by e.g. github.com/David-OConnor/pyflow and github.com/pdm-project/pdm
+__pypackages__/
+
+# Celery stuff
+celerybeat-schedule
+celerybeat.pid
+
+# SageMath parsed files
+*.sage.py
+
+# Environments
+.env
+.venv
+env/
+venv/
+ENV/
+env.bak/
+venv.bak/
+
+# Spyder project settings
+.spyderproject
+.spyproject
+
+# Rope project settings
+.ropeproject
+
+# mkdocs documentation
+/site
+
+# mypy
+.mypy_cache/
+.dmypy.json
+dmypy.json
+
+# Pyre type checker
+.pyre/
+
+# pytype static type analyzer
+.pytype/
+
+# Cython debug symbols
+cython_debug/
+
+# Swift Package Manager
+Packages/
+Package.pins
+Package.resolved
+.build/
+.swiftpm/
+repros/