… cmake/deps.txt
… onnxruntime/core/providers/webgpu/shader_helper.cc
… onnxruntime/core/providers/webgpu/webgpu_context.cc
… onnxruntime/core/providers/webgpu/webgpu_context.h