@guschmue 2025-02-20
the ort web pipeline compiles webgpu ep with emscripten which fails with:
shader_helper.cc:355:45: error: no member named 'ChromiumExperimentalSubgroupMatrix' in 'wgpu::FeatureName'
  355 |   if (device_.HasFeature(wgpu::FeatureName::ChromiumExperimentalSubgroupMatrix))
  
 Possible the headerfile that comes with emscripten doesn't know that featurename yet